remove testConfigureJobAsParameterized and testConfigureSourceCodeByGIT to FreestyleProjectTest
