+++ src/test/java/model/BuildWithParametersPage.java
@@ -40,6 +40,9 @@ public class BuildWithParametersPage extends FreestyleProjectStatusPage {
     @FindBy(xpath = "//a[@href='lastBuild/']")
     private WebElement lastBuildLink;
 
+    @FindBy(linkText = "Configure")
+    private WebElement sideMenuConfigure;
+
     public BuildWithParametersPage clickButtonBuildWithParameters() {
         buttonBuildWithParameters.click();
 
@@ -104,5 +107,11 @@ public class BuildWithParametersPage extends FreestyleProjectStatusPage {
         lastBuildLink.click();
 
         return new StatusPage(getDriver());
+    }
+
+    public FreestyleProjectConfigPage clickConfigureLink() {
+        sideMenuConfigure.click();
+
+        return new FreestyleProjectConfigPage(getDriver());
     }
 }
